Normalize h3 cluster polygons longitudes
